refactor(gateway): extract lifecycle.py from gateway.py — split COMPLETE (6/6)
Final step in the staged gateway.py split.

What moved
- 6 lifecycle helpers extracted into `cognithor/gateway/lifecycle.py`:
    * `start(gw)` — channel + cron startup, signal-handler installation
    * `auto_update_skills(gw)` — daily community-registry sync background task
    * `on_startup_vllm(gw)` — adopt running cognithor-managed vLLM container
    * `on_shutdown_vllm(gw)` — stop container if `auto_stop_on_close` is set
    * `shutdown(gw)` — graceful shutdown with session persistence
    * `rebuild_llm_client(gw, new_backend_type)` — UnifiedLLMClient re-init
      (called from FastAPI `/api/backends/active`)
- Public-API methods (`Gateway.start`, `Gateway.shutdown`,
  `Gateway.on_startup_vllm`, `Gateway.on_shutdown_vllm`,
  `Gateway.rebuild_llm_client`) keep their public signatures via thin
  wrappers in `gateway.py`.
- `register_channel` and `_publish_app_state` STAY in gateway.py per
  the architect blueprint (tightly coupled to `self._channels`).
- `reload_components` STAYS in gateway.py (Admin-API surface).

`gateway.py`: 2 841 -> 2 175 lines (-23 % in this PR; cumulative -67 %
from the pre-split 6 628 LOC across all 6 PRs).

Pre-extraction sweep (Lesson from PRs 1-5)
- Class-level `ClassVar` / `frozenset` / decorators in [1377..2076]: NONE.
- Tests using `inspect.getsource(Gateway.<extracted>)` or
  `monkeypatch.setattr(gateway_module, "<extracted>")`: NONE.
- The cleanest extraction in the entire split. Sweep paid off — zero
  tests needed adapting, zero class-constant restorations needed.

Cumulative split summary (PRs #188..#193)
| Step | Sub-module | gateway.py LOC delta |
|------|------------|----------------------|
| 1/6 message_utils       | 6 628 -> 5 357 (-19%) |
| 2/6 session_mgmt        | 5 357 -> 5 205 (-3%)  |
| 3/6 post_processing     | 5 205 -> 4 825 (-7%)  |
| 4/6 pge_loop            | 4 825 -> 4 249 (-12%) |
| 5/6 message_handler     | 4 249 -> 3 399 (-20%) |
| 6/6 lifecycle           | 3 399 -> 2 175 (-23%) |
| **Final**               | **6 628 -> 2 175 (-67%)** |

Sub-modules created: 7 (message_utils, session_mgmt, post_processing,
pge_loop, message_handler, lifecycle, plus the pre-existing phases/
which the split coordinated with).

What's left in gateway.py (~2 175 LOC orchestrator)
- `Gateway.__init__` (init lifecycle delegation to phases/)
- `Gateway.initialize` (6-phase init coordination)
- `Gateway.cancel_session`, `register_channel`, `_publish_app_state`
- `Gateway.execute_workflow`, `execute_action_plan_as_workflow`,
  `execute_delegation` (Public API)
- `Gateway.reload_components`, `switch_branch` (Admin API)
- `Gateway._sync_core_inventory`, `_sync_core_inventory_legacy`,
  `_record_metric` (utilities)
- 6 wrappers for lifecycle helpers (this PR)
- 6 wrappers for message_handler helpers (PR #192)
- 4 wrappers for post_processing helpers (PR #190)
- 5 wrappers for session_mgmt helpers (PR #189)
- 7 wrappers for message_utils helpers (PR #188)
- 3 wrappers for pge_loop helpers (PR #191)
- Module-level: `MAX_STALLED_MODEL_TURNS`, `advance_stalled_count`,
  `_TOOL_STATUS_KEYS`, `_sanitize_broken_llm_output`,
  `_classify_attachments`, `_build_video_attachment`,
  `_extract_uuid_from_path`, `_PRESEARCH_NO_RESULTS`,
  `_PRESEARCH_NO_ENGINE` (referenced from sub-modules and tests)

Verification
- `pytest tests/test_gateway/ tests/test_core/ tests/test_security/ tests/test_audit_fixes.py tests/test_pattern_documentation.py tests/test_integration/test_observer_flow.py` -> **4 083 passed, 2 skipped**.
- `ruff check + format` -> All checks passed.
- `python -c "from cognithor.gateway.gateway import Gateway; from cognithor.gateway import lifecycle"` -> OK.
